Major reorganization: property_level_analysis.py is greatly reduced and property_store.py's contents have been split between property_providers.py and property_analyzers.py.
